+++ pyonepassword/data/__init__.py
@@ -1,2 +1,5 @@
+from pathlib import Path
+
 TEMPLATE_REGISTRY_JSON = "template-registry.json"
 SVC_ACCOUNT_COMMANDS = "svc_acct_commands"
+OP_VERSION_SUPPORT = Path("op_versions", "version_support.json")